Mini refactoring
