+++ config.py
@@ -1,0 +1,5 @@
+TELEGRAM_TOKEN = '***write here your token***'
+MAPS_KEY = '***write here your key***'
+
+BASE_URL_RADAR = 'https://radar.freedar.uk/VirtualRadar/AircraftList.json'
+BASE_URL_MAPS = 'https://maps.googleapis.com/maps/api/staticmap'
