+++ requirements.txt
@@ -1,4 +1,4 @@
 telepot==12.7
 requests==2.26.0
 Pillow==8.4.0
-googletrans==3.0.0
+googletrans==4.0.0rc1
