+++ utils.py
@@ -1,0 +1,7 @@
+from googletrans import Translator
+
+
+def traduci(sentence, langout):
+    tr = Translator()
+    s = ' ' + tr.translate(sentence, dest=langout).text + ' '
+    return s
